Taylor came over, so stopping work for ripsticking
