--- app/src/main/java/net/conciencia/mensajeandroid/Activitys/HomePageActivity.java
@@ -1,63 +1,0 @@
-package net.conciencia.mensajeandroid.Activitys;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.view.Menu;
-import android.view.MenuItem;
-
-import net.conciencia.mensajeandroid.Constants;
-import net.conciencia.mensajeandroid.R;
-import net.conciencia.mensajeandroid.ContentLoaders.RSSInterface;
-import net.conciencia.mensajeandroid.Fragments.SermonListFragment;
-
-public class HomePageActivity extends AppCompatActivity implements SermonListFragment.SermonListInteraction {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_homepage);
-
-        //Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        //setSupportActionBar(toolbar);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.message_menu, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        Intent intent;
-        switch (item.getItemId()) {
-            case R.id.case_of_the_week_message_menu:
-                intent = new Intent(this, CaseOfTheWeekActivity.class);
-                startActivity(intent);
-                return true;
-            case R.id.info_message_menu:
-                intent = new Intent(this, InformationActivity.class);
-                startActivity(intent);
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-    }
-
-    @Override
-    public void onSermonSelected(RSSInterface rssInterface, int sermonIndex) {
-        //todo
-        //Toast.makeText(this, String.format("Sermon %s (%s) selected", sermon.getTitle(), sermon.getPubDate()), Toast.LENGTH_SHORT).show();
-        Intent intent = new Intent(HomePageActivity.this, SermonViewActivity.class);
-        Bundle extras = new Bundle();
-        extras.putParcelable(Constants.RSS_EXTRA, rssInterface);
-        extras.putInt(Constants.SERMON_INDEX, sermonIndex);
-        intent.putExtras(extras);
-        startActivity(intent);
-    }
-}
